backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
pymysql==1.1.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
minio==7.2.0
pymilvus==2.3.4
langchain==0.0.350
openai==1.3.5
tiktoken==0.5.1
pytest==7.4.3
httpx==0.25.2 